+++ src/main/java/hellfirepvp/astralsorcery/common/crafting/altar/recipes/PrismLensRecipe.java
@@ -77,7 +77,7 @@ public class PrismLensRecipe extends ConstellationRecipe {
     }
 
     @Override
-    public void onCraftClientTick(TileAltar altar, int tick, Random rand) {
+    public void onCraftClientTick(TileAltar altar, long tick, Random rand) {
         super.onCraftClientTick(altar, tick, rand);
 
         Vector3 altarVec = new Vector3(altar);
